macOS: the same signedness fault in five fuzz tests
Found by sweeping rather than by waiting for CI: the macOS build aborted in the
apps, so it never reached the test targets, and these five would have been the
next failure.

All five truncation harnesses build every prefix of a seed with
`seed.begin() + k` where k is size_t. An iterator's difference_type is signed,
so that is a signedness change — the same fault clang reports and GCC does not,
which is why it had never been seen. Widened with an explicit ptrdiff_t cast;
k <= seed.size() throughout, so the cast cannot misrepresent the value.

tests/render/modelparserfuzz.cpp was not in the clang sweep's output yet when
this was written — a grep for the idiom across engine/, tests/ and apps/ found
it, along with four other `begin() + var` sites that are fine because their
variables are already signed (FontFallback, Geometry2D, KdTree2D/3D, editor).

Verified: all five clean under clang with the project's flags, all five still
clean under g++ so Linux does not regress, and the loop shape still produces
every prefix length from 0 to the seed size inclusive.

+++ tests/core/savefilefuzz.cpp
@@ -11,6 +11,7 @@
 #include "maz/core/Checkpoints.hpp"
 #include "maz/io/MoveToFront.hpp"
 
+#include <cstddef>
 #include <cstdint>
 #include <cstdio>
 #include <random>
@@ -38,7 +39,10 @@ static void feedCkpt(const std::vector<std::uint8_t>& b) {
 static void feedMtf(const std::vector<std::uint8_t>& b) { g_sink += io::mtfDecode(b).size(); }
 
 static void truncate(void (*fn)(const std::vector<std::uint8_t>&), const std::vector<std::uint8_t>& seed) {
-    for (std::size_t k = 0; k <= seed.size(); ++k) fn(std::vector<std::uint8_t>(seed.begin(), seed.begin() + k));
+    // seed.begin() + k: k is size_t and an iterator's difference_type is signed, so this is a
+    // signedness change. GCC's -Wconversion lets it pass; clang's implies -Wsign-conversion and
+    // does not, which is why it only ever showed up on macOS. k <= seed.size(), so the cast is safe.
+    for (std::size_t k = 0; k <= seed.size(); ++k) fn(std::vector<std::uint8_t>(seed.begin(), seed.begin() + static_cast<std::ptrdiff_t>(k)));
     std::vector<std::uint8_t> s = seed;
     for (std::size_t i = 0; i < s.size(); ++i) {
         const std::uint8_t orig = s[i];
